functions/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.14.1
psycopg2-binary==2.9.10

# iCal Parsing
icalendar==6.1.1
httpx==0.28.1

# Firebase Push Notifications
firebase-admin==6.6.0

# Scheduler (Cron Jobs)
apscheduler==3.10.4

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Utilities
python-dotenv==1.0.1
pydantic[email-validator,email]==2.10.4
pydantic-settings==2.7.1
Pillow==10.3.0
# Firebase Functions runtime & adapter
firebase-functions>=0.1.0
a2wsgi==1.10.7
